# api/openapi/atlas.openapi.json
{
  "components": {
    "schemas": {
      "Address": {
        "description": "Canonical address and place metadata.",
        "properties": {
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "display": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display"
          },
          "full_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Address"
          },
          "geo_specificity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geo Specificity"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          }
        },
        "title": "Address",
        "type": "object"
      },
      "ContactInfo": {
        "description": "Canonical public contact surface.",
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "social_media": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Social Media"
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website"
          }
        },
        "title": "ContactInfo",
        "type": "object"
      },
      "CoverageCount": {
        "description": "Issue coverage count.",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "issue_area_id": {
            "title": "Issue Area Id",
            "type": "string"
          }
        },
        "required": [
          "issue_area_id",
          "count"
        ],
        "title": "CoverageCount",
        "type": "object"
      },
...
      "EntityCollectionResponse": {
        "description": "Collection response for entities.",
        "properties": {
          "facets": {
            "anyOf": [
              {
                "additionalProperties": {
                  "items": {
                    "$ref": "#/components/schemas/FacetOption"
                  },
                  "type": "array"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facets"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/EntityResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "place": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Address"
              },
              {
                "type": "null"
              }
            ]
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total"
        ],
        "title": "EntityCollectionResponse",
        "type": "object"
      },
...
      "EntityResponse": {
        "description": "Canonical public entity shape.",
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "affiliated_org_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Affiliated Org Id"
          },
          "contact": {
            "$ref": "#/components/schemas/ContactInfo"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "flag_summary": {
            "$ref": "#/components/schemas/FlagSummary"
          },
          "freshness": {
            "$ref": "#/components/schemas/FreshnessInfo"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "issue_area_ids": {
            "items": {
              "type": "string"
            },
            "title": "Issue Area Ids",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "resource_uri": {
            "title": "Resource Uri",
            "type": "string"
          },
          "source_count": {
            "default": 0,
            "title": "Source Count",
            "type": "integer"
          },
          "source_types": {
            "items": {
              "type": "string"
            },
            "title": "Source Types",
            "type": "array"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "verified": {
            "title": "Verified",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "type",
          "name",
          "description",
          "address",
          "contact",
          "active",
          "verified",
          "freshness",
          "created_at",
          "updated_at",
          "resource_uri"
        ],
        "title": "EntityResponse",
        "type": "object"
      },
...
      "FacetOption": {
        "description": "A single filter option with result count.",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "value",
          "count"
        ],
        "title": "FacetOption",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IssueSignalsResponse": {
        "description": "Issue signal response for a place.",
        "properties": {
          "issues": {
            "items": {
              "$ref": "#/components/schemas/IssueSignalSummary"
            },
            "title": "Issues",
            "type": "array"
          },
          "place": {
            "$ref": "#/components/schemas/Address"
          },
          "resource_uri": {
            "title": "Resource Uri",
            "type": "string"
          }
        },
        "required": [
          "place",
          "resource_uri"
        ],
        "title": "IssueSignalsResponse",
        "type": "object"
      },
      "PlaceCoverageResponse": {
        "description": "Coverage summary for a place.",
        "properties": {
          "covered_issue_area_ids": {
            "items": {
              "type": "string"
            },
            "title": "Covered Issue Area Ids",
            "type": "array"
          },
          "entity_count": {
            "title": "Entity Count",
            "type": "integer"
          },
          "issue_counts": {
            "items": {
              "$ref": "#/components/schemas/CoverageCount"
            },
            "title": "Issue Counts",
            "type": "array"
          },
          "missing_issue_area_ids": {
            "items": {
              "type": "string"
            },
            "title": "Missing Issue Area Ids",
            "type": "array"
          },
          "place": {
            "$ref": "#/components/schemas/Address"
          },
          "resource_uri": {
            "title": "Resource Uri",
            "type": "string"
          },
          "thin_issue_area_ids": {
            "items": {
              "type": "string"
            },
            "title": "Thin Issue Area Ids",
            "type": "array"
          },
          "uncovered_domains": {
            "items": {
              "type": "string"
            },
            "title": "Uncovered Domains",
            "type": "array"
          }
        },
        "required": [
          "place",
          "entity_count",
          "resource_uri"
        ],
        "title": "PlaceCoverageResponse",
        "type": "object"
      },
      "PlaceIdentityResponse": {
        "description": "REST or MCP representation of a place resource.",
        "properties": {
          "place": {
            "$ref": "#/components/schemas/Address"
          },
          "resource_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Uri"
          }
        },
        "required": [
          "place"
        ],
        "title": "PlaceIdentityResponse",
        "type": "object"
      },
      "PlaceProfileResponse": {
        "description": "Structured place profile with flexible data blocks.",
        "properties": {
          "demographics": {
            "additionalProperties": true,
            "title": "Demographics",
            "type": "object"
          },
          "economics": {
            "additionalProperties": true,
            "title": "Economics",
            "type": "object"
          },
          "education": {
            "additionalProperties": true,
            "title": "Education",
            "type": "object"
          },
          "health": {
            "additionalProperties": true,
            "title": "Health",
            "type": "object"
          },
          "housing": {
            "additionalProperties": true,
            "title": "Housing",
            "type": "object"
          },
          "place": {
            "$ref": "#/components/schemas/Address"
          },
          "provenance": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Provenance",
            "type": "array"
          },
          "resource_uri": {
            "title": "Resource Uri",
            "type": "string"
          }
        },
        "required": [
          "place",
          "resource_uri"
        ],
        "title": "PlaceProfileResponse",
        "type": "object"
      },
...
      "SourceCollectionResponse": {
        "description": "Collection response for sources.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SourceResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "place": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Address"
              },
              {
                "type": "null"
              }
            ]
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total"
        ],
        "title": "SourceCollectionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "email": "atlas@example.com",
      "name": "Rebuilding America",
      "url": "https://atlas.example.com/"
    },
    "description": "Atlas exposes place-first civic research data through a typed REST API. The public contract includes entities, sources, issue areas, domains, place profiles, issue signals, discovery runs, and anonymous data flags.",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "summary": "Structured civic research data for Atlas places, entities, sources, and discovery runs.",
    "title": "Atlas REST API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/places/{place_key}": {
      "get": {
        "description": "Return the canonical Atlas identity for a normalized place key.",
        "operationId": "getPlace",
        "parameters": [
          {
            "in": "path",
            "name": "place_key",
            "required": true,
            "schema": {
              "title": "Place Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceIdentityResponse"
                }
              }
            },
            "description": "The canonical Atlas place resource."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a place",
        "tags": [
          "places"
        ]
      }
    },
    "/api/places/{place_key}/coverage": {
      "get": {
        "description": "Return Atlas coverage counts and gaps for one place.",
        "operationId": "getPlaceCoverage",
        "parameters": [
          {
            "in": "path",
            "name": "place_key",
            "required": true,
            "schema": {
              "title": "Place Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "issue_area",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Issue Area"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceCoverageResponse"
                }
              }
            },
            "description": "Atlas coverage metadata for the requested place."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get place coverage",
        "tags": [
          "places"
        ]
      }
    },
    "/api/places/{place_key}/entities": {
      "get": {
        "description": "List Atlas entities associated with one place, with optional issue-area, entity-type, source-type, and text filters.",
        "operationId": "listPlaceEntities",
        "parameters": [
          {
            "in": "path",
            "name": "place_key",
            "required": true,
            "schema": {
              "title": "Place Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "issue_area",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Issue Area"
            }
          },
          {
            "in": "query",
            "name": "entity_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Type"
            }
          },
          {
            "in": "query",
            "name": "source_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Type"
            }
          },
          {
            "in": "query",
            "name": "text",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Text"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityCollectionResponse"
                }
              }
            },
            "description": "A paginated collection of Atlas entities for the requested place."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List place entities",
        "tags": [
          "places"
        ]
      }
    },
    "/api/places/{place_key}/issue-signals": {
      "get": {
        "description": "Return structured issue-area signals for one place based on Atlas entities and sources.",
        "operationId": "getPlaceIssueSignals",
        "parameters": [
          {
            "in": "path",
            "name": "place_key",
            "required": true,
            "schema": {
              "title": "Place Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "issue_area",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Issue Area"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueSignalsResponse"
                }
              }
            },
            "description": "Issue-area signal summaries for the requested place."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get place issue signals",
        "tags": [
          "places"
        ]
      }
    },
    "/api/places/{place_key}/profile": {
      "get": {
        "description": "Return structured demographic and socioeconomic context for one place.",
        "operationId": "getPlaceProfile",
        "parameters": [
          {
            "in": "path",
            "name": "place_key",
            "required": true,
            "schema": {
              "title": "Place Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceProfileResponse"
                }
              }
            },
            "description": "The Atlas place profile for the requested place."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a place profile",
        "tags": [
          "places"
        ]
      }
    },
    "/api/places/{place_key}/sources": {
      "get": {
        "description": "List source records associated with one place, with optional issue-area, source-type, and text filters.",
        "operationId": "listPlaceSources",
        "parameters": [
          {
            "in": "path",
            "name": "place_key",
            "required": true,
            "schema": {
              "title": "Place Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "issue_area",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Issue Area"
            }
          },
          {
            "in": "query",
            "name": "source_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Type"
            }
          },
          {
            "in": "query",
            "name": "text",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Text"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceCollectionResponse"
                }
              }
            },
            "description": "A paginated collection of Atlas sources for the requested place."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List place sources",
        "tags": [
          "places"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://atlas.example.com"
    },
    {
      "url": "http://localhost:8000"
    },
    {
      "url": "/"
    }
  ],
...
}